+++ assignment5/app/src/main/java/com/wafflestudio/assignment5/Clock.kt
@@ -1,0 +1,11 @@
+package com.wafflestudio.assignment5
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun Clock(
+    modifier: Modifier = Modifier
+){
+
+}
