+++ assignment5/app/src/main/java/com/wafflestudio/assignment5/Movie.kt
@@ -1,0 +1,18 @@
+package com.wafflestudio.assignment5
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun Movie(
+    modifier: Modifier = Modifier
+){
+    Column(
+        modifier = modifier.fillMaxWidth()
+    ){
+        LazyColumn(modifier = modifier){
+        }
+    }
+}
